backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
supabase
httpx
openai
google-generativeai
numpy
pandas
pytest
pytest-asyncio
reportlab
